third_party/blink/web_tests/flag-specific/highdpi/fast/css-grid-layout/grid-auto-repeat-huge-grid-004-expected.txt

This is a testharness.js-based test.
FAIL Test that we don't get more than kGridMaxTracks repetitions even on very wide grids with gaps and min-width. assert_equals: expected "20px" but got "19.9896px"
FAIL Test that we don't get more than kGridMaxTracks repetitions even on very tall grids with gaps and min-height. assert_equals: expected "20px" but got "19.9896px"
Harness: the test ran to completion.
